requirements.txt
requests==2.32.3
gspread==6.1.4
google-auth==2.34.0
openai>=1.51.0
